renamed adapter to Criteria
